cohort-1/track-3/edupulse/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
google-genai==1.14.0
pytest==8.3.4
